Move prometheus files to runtime
In the current implementation, we are using prometheus to export metrics
from the single/multi/ssh deployer, which are part of the weaver repo.

However, in the upcoming kubernetes deployer (that's in a different
repo), we will rely on prometheus to export the metrics as well.
However, we can't (shouldn't) depend on internal files from /weaver.

Given that the prometheus metric translator will be used in other repos,
we should move it to runtime/prometheus (we already have
runtime/perfetto).
